[FEATURE] Integrate jQuery UI 1.11.2
jQuery UI 1.11 has support for AMD allowing
to not load the whole jQuery UI library.

This version is now included and used
in favor of jQuery UI 1.10 and scriptaculous.

Furthermore only Core and Interaction components
that are necessary to replace ExtJS and
Scriptaculous functionality are included, no
widgets.

Existing backend modules are encouraged to use
the RequireJS API and to specifically require
only the components needed.

Resolves: #62916
Releases: master
Change-Id: Ib4404704e96778054bf06dfe679522d6dc5bb621
Reviewed-on: http://review.typo3.org/34060
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>