[BUGFIX] Tasks Center section menu is hardcoded
Since the refactoring of the tasks center in 4.5, the section menu is
hardcoded to information and tasks and cannot be changed.

This is because the menu items are collected inside the array key 'mode'
instead of 'function'.
Because 'function' is hardcoded in the logic of BaseScriptClass, copy
the four lines from BaseScriptClass instead of calling
parent:menu_config() in the TaskModuleController.

Additionally, 'Tasks' is hardcoded as a label. Replace it with a
localized value.

Note: 'function' is preserved for backward compatibility and is already
in use for the main tasks logic.

Use page.mod.user_task.menu.mode.information = 0 to blind an individual
item, e.g. information.

Some further code cleanups are included.

Resolves: #63529
Releases: master
Change-Id: Ic2666dde7e794689011d8bac4996cda4216988c8
Reviewed-on: http://review.typo3.org/34947
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>